Moving register module codec method to utilities/module/
